pyproject.toml
[tool.poetry]
name = "tmpmail"
version = "0.1.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]

[tool.poetry.dependencies]
python = "^3.10"
salmon-mail = "^3.2.0"
peewee = "^3.14.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
